\name{ReproductiveBurden-package}
\alias{ReproductiveBurden-package}
\alias{ReproductiveBurden}
\docType{package}
\title{
\packageTitle{ReproductiveBurden}
}
\description{
\packageDescription{ReproductiveBurden}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ReproductiveBurden}
\packageIndices{ReproductiveBurden}
The simRA function from the ReproductiveBurden-package simulates and graphically illustrates the expected reproductive burden during pregnancy for a live-bearing species as a function of time and for a specific degree of superfetation, litter size, egg mass at fertilization, and offspring mass at birth.
}
\author{
\packageAuthor{ReproductiveBurden}

Maintainer: \packageMaintainer{ReproductiveBurden}
}
\references{
Hagmayer, A., Furness, A. I., Reznick, D. N., and Pollux, B. J. (2018). Maternal size and body condition predict the amount of post-fertilization maternal provisioning in matrotrophic fish. Ecology and Evolution, 8(24):12386-12396
}

\keyword{ package }
\seealso{

}
\examples{
simRA(SF = 3,LS = 4,numberLitter = 5,eggWeight = 1,offspringWeight = 8)
}
